% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burndownChart.R
\name{burndownChart}
\alias{burndownChart}
\title{Wrapper function for burndown plotly charts using Reynold's standard}
\usage{
burndownChart(
  burndownBars,
  backlogLine,
  legendHeight = 1.05,
  horizontalLegend = NULL
)
}
\arguments{
\item{burndownBars}{is a data table with the bar values; requires the columns Period, lens, and value. Note that the variable column in the example is not required; a column to distinguish incoming/outgoing items is likely helpful (and a prerequisite for any aggregation done prior).}

\item{backlogLine}{is a data table with the backlog values for each period; requres the columns Period and Backlog.}

\item{horizontalLegend}{is a boolean value indicating whether the legend should be shown horizontally at the top or vertically on the right.}
}
\description{
Wrapper function for burndown plotly charts using Reynold's standard
}
\examples{
set.seed(1)
months <- c("Jan-23", "Feb-23", "Mar-23",  "Apr-23", "May-23")
data   <- rbind(data.table(Period   = factor(months, levels = months),
                           lens     = "Bucket 1",
                           variable = "Incoming",
                           value    = sample(0:30, 5, replace = FALSE)),
                data.table(Period   = factor(months, levels = months),
                           lens     = "Bucket 2",
                           variable = "Incoming",
                           value    = sample(0:30, 5, replace = FALSE)),
                data.table(Period   = factor(months, levels = months),
                           lens     = "Bucket 1",
                           variable = "Outgoing",
                           value    = -sample(0:30, 5, replace = FALSE)),
                data.table(Period   = factor(months, levels = months),
                           lens     = "Bucket 2",
                           variable = "Outgoing",
                           value    = -sample(0:30, 5, replace = FALSE)))
priorQueue <- 50
totalLine  <- data[, .(Backlog = priorQueue + sum(value)), by = .(Period)]
burndownChart(burndownBars  = data,
              backlogLine  = totalLine)
}
\keyword{burndown,}
\keyword{plot}
\keyword{plotly,}
