% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burndownStandard.R
\name{burndownStandard}
\alias{burndownStandard}
\title{Standard for burndown charts}
\usage{
burndownStandard(
  input,
  output,
  session,
  burndownData,
  keyword,
  boxTitle = "Burndown",
  burndownActionButtonGroupList = NULL,
  actionButtonInputFunction = NULL
)
}
\arguments{
\item{input}{is passed from Shiny server}

\item{output}{is passed from Shiny server}

\item{session}{is passed from Shiny server}
}
\description{
Standard for burndown charts
}
\keyword{filters}
