% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliToken.R
\name{cliToken}
\alias{cliToken}
\title{Get CLI token}
\usage{
cliToken(resource)
}
\arguments{
\item{resource}{(e.g., https://storage.azure.com)}
}
\description{
Get CLI token
}
\examples{
cliToken(resource = "https://storage.azure.com")

}
\keyword{cli,}
\keyword{token}
