% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateMapping.R
\name{dateMapping}
\alias{dateMapping}
\title{SQL query for day, week, month formatting from dim.Date}
\usage{
dateMapping(startDate, endDate, moadConn = NULL)
}
\arguments{
\item{startDate}{should be a date of type character or Date in yyyy-mm-dd format}

\item{endDate}{should be a date of type character or Date in yyyy-mm-dd format}

\item{moadConn}{should be an ODBC connection to MoAD; if not provided, it will connect to MoAD prod by default}
}
\description{
SQL query for day, week, month formatting from dim.Date
}
\examples{
chieR::dateMapping(startDate = as.Date("2023-01-01"), endDate = as.Date("2023-01-31"))

}
\keyword{date}
