% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_user_details.R
\name{fetch_user_email}
\alias{fetch_user_email}
\title{Function to fetch user details}
\usage{
fetch_user_email(access_token)
}
\description{
Function to fetch user details
}
\keyword{email}
\keyword{token,}
\keyword{user,}
