% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Function to get token using authorization code}
\usage{
get_token(code, app, tenant, password)
}
\description{
Function to get token using authorization code
}
\keyword{token}
