% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAzureStorageBlob.R
\name{getAzureStorageBlob}
\alias{getAzureStorageBlob}
\title{Get blob from Azure Storage}
\usage{
getAzureStorageBlob(storageAccount, accountKey, container, filename)
}
\arguments{
\item{storageAccount}{Storage account name}

\item{accountKey}{Account key}

\item{container}{Container}

\item{filename}{File name}
}
\description{
Get blob from Azure Storage
}
\examples{
getAzureStorageBlob()
}
\keyword{storage}
