% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColors.R
\name{getColors}
\alias{getColors}
\title{Input a string indicating the type of color palette you want.
Options are Standard, Gen, HW Bucket, and Capability Gap}
\usage{
getColors(type = "Standard")
}
\arguments{
\item{type}{should be "Standard", "Gen", "HW Bucket", or "Capability Gap"}
}
\description{
Input a string indicating the type of color palette you want.
Options are Standard, Gen, HW Bucket, and Capability Gap
}
\examples{
getColors()
}
\keyword{colors}
