% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFontSize.R
\name{getFontSize}
\alias{getFontSize}
\title{Get font size}
\usage{
getFontSize(type)
}
\arguments{
\item{type}{should be "axisTitle", "axis", "legend", or "annotation"}
}
\description{
Get font size
}
\examples{
getFontSize()
}
\keyword{font}
\keyword{size}
