% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miToken.R
\name{miToken}
\alias{miToken}
\title{Get managed identity token}
\usage{
miToken(resource)
}
\arguments{
\item{resource}{(e.g., https://storage.azure.com)}
}
\description{
Get managed identity token
}
\examples{
miToken(resource = "https://storage.azure.com")

}
\keyword{identity,}
\keyword{managed,}
\keyword{token}
