% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeWidgets.R
\name{organizeWidgets}
\alias{organizeWidgets}
\title{Show automated filters, fixed widgets (e.g., for time interval, date range, lens) based on the standard}
\usage{
organizeWidgets(filters, dateLensWidgets, maxrow1filters = 3)
}
\arguments{
\item{filters}{are the automated widgets}

\item{dateLensWidgets}{are the date/lens-related widgets}
}
\description{
Show automated filters, fixed widgets (e.g., for time interval, date range, lens) based on the standard
}
\keyword{filters}
