% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyChartWrapper.R
\name{plotlyChartWrapper}
\alias{plotlyChartWrapper}
\title{Input a string indicating the type of color palette you want.
Options are Standard, Gen, HW Bucket, and Capability Gap}
\usage{
plotlyChartWrapper(
  data,
  mode,
  barmode = "relative",
  colorPalette = "Standard",
  yaxisLabel = "",
  showBarTotals = FALSE,
  barTotalDecimal = 0,
  percent = FALSE,
  legendHeight = 1.1
)
}
\arguments{
\item{mode}{should be "bar" or "lines"}

\item{barmode}{should be options like "group", "relative", or "stacked"}

\item{colorPalette}{should be acceptable types for chieR::getColor()}
}
\description{
Input a string indicating the type of color palette you want.
Options are Standard, Gen, HW Bucket, and Capability Gap
}
\examples{
data <- data.table(WorldPhones)[, Period := 2001:2007][, Order := 1:7][]
data <- melt(data, c("Period", "Order"), variable.name = "lens")

plotlyChartWrapper(data,
                   mode            = "lines",
                   yaxisLabel      = "Number of telephones")

plotlyChartWrapper(data,
                   mode            = "bar",
                   barmode         = "relative",
                   yaxisLabel      = "Number of telephones",
                   showBarTotals   = TRUE)

plotlyChartWrapper(data,
                   mode            = "bar",
                   barmode         = "group",
                   yaxisLabel      = "Number of telephones")
plotlyChartWrapper()
}
\keyword{bar,}
\keyword{line,}
\keyword{plot}
\keyword{plotly,}
