% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyLayout.R
\name{plotlyLayout}
\alias{plotlyLayout}
\title{Plotly defaults that conform to CHIE standard}
\usage{
plotlyLayout(
  plot,
  xaxisLabel = "",
  yaxisLabel = "",
  horizontalLegend = TRUE,
  legendHeight = 1
)
}
\arguments{
\item{plot}{is a plotly object}

\item{xaxisLabel}{is a string}

\item{yaxisLabel}{is a string}

\item{horizontalLegend}{Boolean for whether the legend should be shown at the top or on the right}

\item{legendHeight}{Legends at the top have height issues, so set this to some appropriate value > 1 (e.g., 1.25)}
}
\description{
Plotly defaults that conform to CHIE standard
}
\examples{
plotlyLayout()
}
\keyword{plotly}
