% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictabilityChart.R
\name{predictabilityChart}
\alias{predictabilityChart}
\title{Wrapper function for predictability plotly charts using Reynold's standard}
\usage{
predictabilityChart(
  data,
  minBarLabel = 1,
  barAxisHeight = NULL,
  predictabilityLineName = "Predictability",
  sourceName = NULL,
  legendHeight = NULL
)
}
\arguments{
\item{data}{should be a data table in wide format with the columns Period and Predictability, and at least one other column that would be represented with bars.}

\item{minBarLabel}{is the minimum value for which bar labels should be shown; assumed to be 1, but this can be crowded.}

\item{barAxisHeight}{Height of axis for bars; optional.}

\item{predictabilityLineName}{Name of line as shown in legend. Assumed by default to be Predictability.}

\item{sourceName}{source to be passed to plot_ly() in preparation for linking with event_data()}

\item{legendHeight}{Adjust the height of the legend if necessary}
}
\description{
Wrapper function for predictability plotly charts using Reynold's standard
}
\examples{
set.seed(1)
periods <- c("Jan-23", "Feb-23", "Mar-23", "Apr-23", "May-23", "Jun-23")
data <- data.table(Period  = factor(periods, levels = periods),
                   `Bar 1` = round(runif(length(periods), 1, 15)),
                   `Bar 2` = round(runif(length(periods), 1, 15)),
                   Predictability = runif(length(periods), 1, 100) / 100)
predictabilityChart(data                   = copy(data),
                    predictabilityLineName = "Cycle Time")
}
\keyword{plot}
\keyword{plotly,}
\keyword{predictability,}
