% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semesterDates.R
\name{semesterDates}
\alias{semesterDates}
\title{Wrapper function to get all of the dates associated with the semester for a particular date or semester name}
\usage{
semesterDates(semesterName = NULL, date = Sys.Date(), moadConn = NULL)
}
\arguments{
\item{semesterName}{should be a semester name like "Ga" or "Zn". If no value is provided, semester information will be returned based on the date.}

\item{date}{should be a date of type character or Date in yyyy-mm-dd format. If semesterName is not provided, the function will return a data table with the dates associated with the semester relevant to the date provided.}

\item{moadConn}{should be an ODBC connection to MoAD; if not provided, it will connect to MoAD prod by default}
}
\description{
Wrapper function to get all of the dates associated with the semester for a particular date or semester name
}
\examples{
chieR::dateMapping(startDate = as.Date("2023-01-01"), endDate = as.Date("2023-01-31"))

}
\keyword{date}
\keyword{semester,}
