% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlConnect.R
\name{sqlConnect}
\alias{sqlConnect}
\title{Connect to MoAD, other databases.}
\usage{
sqlConnect(server)
}
\arguments{
\item{server}{should be "chiemoaddev" or "dsireport"}
}
\description{
Connect to MoAD, other databases.
}
\examples{
sqlConnect()
}
\keyword{connection}
