% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardInputs.R
\name{standardInputs}
\alias{standardInputs}
\title{Standard for agility charts}
\usage{
standardInputs(
  data,
  filterCols,
  type,
  naToUnknownTF = FALSE,
  moadConn = chieR::sqlConnect(server = "chiemoadprd"),
  semesterDates = chieR::semesterDates(moadConn = moadConn)[]
)
}
\arguments{
\item{type}{must be agility, trend, burndown, or throughput}
}
\description{
Standard for agility charts
}
\keyword{inputs}
\keyword{standard,}
\keyword{wrapper,}
