% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throughputStandard.R
\name{throughputStandard}
\alias{throughputStandard}
\title{Standard for throughput charts}
\usage{
throughputStandard(
  input,
  output,
  session,
  throughputData,
  keyword,
  boxTitle = "Throughput",
  throughputActionButtonGroupList = NULL,
  actionButtonInputFunction = NULL,
  cumulateBySemester = TRUE,
  filterDefaultSelectedValues = NULL,
  defaultLens = "None"
)
}
\arguments{
\item{input}{is passed from Shiny server}

\item{output}{is passed from Shiny server}

\item{session}{is passed from Shiny server}
}
\description{
Standard for throughput charts
}
\keyword{filters}
