% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendStandard.R
\name{trendStandard}
\alias{trendStandard}
\title{Standard for trend charts}
\usage{
trendStandard(
  input,
  output,
  session,
  trendData,
  keyword,
  title,
  percentTF = FALSE,
  timeIntervalChoices = c("Day", "Week", "Month"),
  defaultTimeInterval = "Month",
  addlTableColumns = NULL,
  showTable = FALSE,
  averageByPeriod = FALSE,
  lastDayInPeriod = FALSE,
  normalizeBynRow = FALSE,
  filterDefaultSelectedValues = NULL,
  defaultLens = "None"
)
}
\arguments{
\item{input}{is passed from Shiny server}

\item{output}{is passed from Shiny server}

\item{session}{is passed from Shiny server}
}
\description{
Standard for trend charts
}
\keyword{filters}
