% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vline.R
\name{vline}
\alias{vline}
\title{Wrapper function for creating vertical lines with plotly::layout() shapes}
\usage{
vline(x)
}
\arguments{
\item{x}{is the x-axis location of the vertical line.}
}
\description{
Wrapper function for creating vertical lines with plotly::layout() shapes
}
\keyword{line}
\keyword{vertical,}
\keyword{vline,}
