% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agilityStandard.R
\name{agilityStandard}
\alias{agilityStandard}
\title{Standard for agility charts}
\usage{
agilityStandard(
  input,
  output,
  session,
  agilityData,
  keyword,
  boxTitle = "Agility",
  agilityActionButtonGroupList = NULL,
  actionButtonInputFunction = NULL,
  filterDefaultSelectedValues = NULL,
  showTable = FALSE,
  addlTableColumns = NULL
)
}
\arguments{
\item{input}{is passed from Shiny server}

\item{output}{is passed from Shiny server}

\item{session}{is passed from Shiny server}
}
\description{
Standard for agility charts
}
\keyword{filters}
