% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naToUnknown.R
\name{naToUnknown}
\alias{naToUnknown}
\title{Wrapper function to change NA values to "Unknown". Important because filtering will ignore NA values.}
\usage{
naToUnknown(data, filterCols)
}
\arguments{
\item{data}{is a data table}

\item{filterCols}{is a vector of column names in data}
}
\description{
Wrapper function to change NA values to "Unknown". Important because filtering will ignore NA values.
}
\keyword{NA,}
\keyword{Unknown}
\keyword{filter,}
