% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naToUnknown.R
\name{naToUnknown}
\alias{naToUnknown}
\title{Wrapper function to change NA values to "Unknown". Important because filtering will ignore NA values.}
\usage{
naToUnknown(data, filterCols)
}
\arguments{
\item{x}{is the x-axis location of the vertical line.}
}
\description{
Wrapper function to change NA values to "Unknown". Important because filtering will ignore NA values.
}
\keyword{line}
\keyword{vertical,}
\keyword{vline,}
