% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyChartWrapper.R
\name{plotlyChartWrapper}
\alias{plotlyChartWrapper}
\title{Wrapper function for line/bar plotly charts using Reynold's standard}
\usage{
plotlyChartWrapper(
  data,
  mode,
  barmode = "relative",
  colorPalette = "Standard",
  yaxisLabel = "",
  showBarTotals = FALSE,
  barTotalDecimal = 0,
  percent = FALSE,
  legendHeight = 1.1
)
}
\arguments{
\item{data}{should be a data table in long format (i.e., melted; see example for details) with the columns Period, lens (if there's no lens, set lens to be the same value for every row), and value. The Order column is optional; if it exists, it should be unique to each Period value.}

\item{mode}{should be "bar" or "lines"}

\item{barmode}{should be options like "group", "relative", or "stacked"}

\item{colorPalette}{should be acceptable types for chieR::getColor()}
}
\description{
Wrapper function for line/bar plotly charts using Reynold's standard
}
\examples{
data <- data.table(WorldPhones)[, Period := 2001:2007][, Order := 1:7][]
data <- melt(data, c("Period", "Order"), variable.name = "lens")

plotlyChartWrapper(data,
                   mode            = "lines",
                   yaxisLabel      = "Number of telephones")

plotlyChartWrapper(data,
                   mode            = "bar",
                   barmode         = "relative",
                   yaxisLabel      = "Number of telephones",
                   showBarTotals   = TRUE)

plotlyChartWrapper(data,
                   mode            = "bar",
                   barmode         = "group",
                   yaxisLabel      = "Number of telephones")
}
\keyword{bar,}
\keyword{line,}
\keyword{plot}
\keyword{plotly,}
