% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlConnect.R
\name{sqlConnect}
\alias{sqlConnect}
\title{Connect to MoAD, other databases.}
\usage{
sqlConnect(
  server,
  token = NULL,
  managedID = NULL,
  clientID = NULL,
  clientSecret = NULL
)
}
\arguments{
\item{server}{should be "chiemoaddev", "chiemoadprd", or "dsireport"}
}
\description{
Connect to MoAD, other databases.
}
\examples{
sqlConnect()
}
\keyword{connection}
